requirements.txt
Babel==2.6.0
boto3==1.7.62
botocore==1.10.62
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.5
django-allauth==0.36.0
django-appconf==1.0.2
django-compressor==2.2
django-phonenumber-field==2.0.0
django-storages==1.6.6
django-widget-tweaks==1.4.2
docutils==0.14
gunicorn==19.9.0
idna==2.7
jmespath==0.9.3
oauthlib==2.1.0
phonenumberslite==8.9.10
psycopg2==2.7.5
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.5
rcssmin==1.0.6
requests==2.19.1
requests-oauthlib==1.0.0
rjsmin==1.0.12
s3transfer==0.1.13
six==1.11.0
urllib3==1.23
whitenoise==3.3.1
